execore: clean UX for another email error path
If you provide an invalid email address,
this is the response that Postmark returns.
(Yes, string matching sadness.)
Clean it up a bit before showing the user.
The Error call will probably end up being too noisy in #page,
but we can tune that later when it starts to hurt.
In the meantime, I'd rather know than not,
because our local email verification should more
or less prevent this from happening.