execore, exedb: add a per-user VM createion disabled bit
In the future, I am going to regret the simplicity of this,
and wish it were a max VM count, or somehow plugged into
some sort of billing DB, or at least had a text reason.
But we need this for spam management as part of launch,
so I am taking the tech debt. The future can blame me.